Verify all packages except app-code are reproducible in CI
With the builder refactor, we can now run two sets of builds in CI
without it being too slow: at ~14m it finishes before the full app-test
job. And this will get dramatically faster once we have reproducible
wheels.

We run the main SecureDrop and OSSEC package builds twice and then
diffoscope all the debs (except securedrop-app-code). Using reprotest
isn't super useful here since the container would hide many of the
intended changes. But it's a nice TODO for the future. In that sense,
this is mostly proving that in a controlled environment, the builds are
reproducible.